+++ flask_via/examples/blueprints/baz/routes.py
@@ -1,15 +1,16 @@
 # -*- coding: utf-8 -*-
 
 """
-flask_via.examples.blueprints.foo.routes
+flask_via.examples.blueprints.baz.routes
 ========================================
 
 A blueprint ``Flask-Via`` example Flask application.
 """
 
-from flask_via.examples.blueprints.foo import views
+from flask_via.examples.blueprints.baz import views
 from flask.ext.via.routers import default
 
+
 routes = [
-    default.Functional('/baz', views.baz),
+    default.Functional('/', views.baz),
 ]